added few tests

+++ contracts/Main.sol
@@ -9,67 +9,66 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.8;
 
-contract VoteWeb3{
-
+contract VoteWeb3 {
     // MANAGER SHIT - BOX STARTS
     address private manager;
 
-    constructor(){
+    constructor() {
         // WHOEVER DEPLOYS CONTRACT - IS THE MANAGER OF THE DAPP
         manager = msg.sender;
     }
 
     // function to get manager
-    function getManager() public view returns(address){
+    function getManager() public view returns (address) {
         return manager;
     }
 
     // STAY OUT OF THE ABOVE BOX
 
     // Election Struct
-    struct Election{
-        uint electionIndex;
+    struct Election {
+        uint256 electionIndex;
         Candidate[] approvedcandidates;
         Candidate[] requestedCandidates;
-        uint startDateAndTime;
-        uint endDateAndTime;
-        uint totalVotes;
-        uint winnerVotes;
+        uint256 startDateAndTime;
+        uint256 endDateAndTime;
+        uint256 totalVotes;
+        uint256 winnerVotes;
         address winner;
         ElectionState electionState;
     }
 
     // candidateVotes[0][0x5B38Da6a701c568545dCfcB03FcB875f56beddC4] => will return the votes of given address in the given indexed election
-    mapping(uint => mapping(address => uint)) candidateVotes;
+    mapping(uint256 => mapping(address => uint256)) candidateVotes;
 
     // to know if candidate already exist in a given election
-    mapping(uint => mapping(address => bool)) candidateExist;
+    mapping(uint256 => mapping(address => bool)) candidateExist;
 
     // to know if candidate already requested for a given election
-    mapping(uint => mapping(address => bool)) candidateRequested;
+    mapping(uint256 => mapping(address => bool)) candidateRequested;
 
     // to get candidate to add after request - values will be set while requesting
-    mapping(uint => mapping(address => Candidate)) getCandidateInfo;
+    mapping(uint256 => mapping(address => Candidate)) getCandidateInfo;
 
     // to know if user has already voted or not
-    mapping(uint => mapping(address => bool)) userVoted;
+    mapping(uint256 => mapping(address => bool)) userVoted;
 
     // Candidate Struct
-    struct Candidate{
+    struct Candidate {
         address candidateAddress;
         string candidateName;
         Symbol candidateSymbol;
     }
 
     // Symbols to choose from - passed as uint
-    enum Symbol{
+    enum Symbol {
         BJP,
         AAP,
         CON
     }
 
     // States of Election to perform functionalities
-    enum ElectionState{
+    enum ElectionState {
         Created,
         Ongoing,
         Finished,
@@ -77,7 +76,7 @@ contract VoteWeb3{
     }
 
     // winner votes - this will keep getting updated as people votes
-    uint highestVotes = 0;
+    uint256 highestVotes = 0;
 
     // winner address - this will keep getting updated as people votes
     address currentWinningCandidate;
@@ -89,7 +88,7 @@ contract VoteWeb3{
     Election electioInstance;
 
     // to manage election indexes - to retrive elections - IMP variable
-    uint currentElectionIndex = 0;
+    uint256 currentElectionIndex = 0;
 
     // Events to emit on function calls
 
@@ -97,16 +96,16 @@ contract VoteWeb3{
 
     event GetElection(Election);
 
-    event ElectionStarted(Election); 
-    
-    event ElectionEnded(Election); 
+    event ElectionStarted(Election);
+
+    event ElectionEnded(Election);
 
     event CandidateRequested(Candidate);
 
     event CandidateAdded(Candidate);
-    
+
     // creating new election
-    function createElection(uint startTime, uint endTime) public {
+    function createElection(uint256 startTime, uint256 endTime) public {
         // msg.sender should be manager
         require(msg.sender == manager, "You are not Authorized");
 
@@ -118,7 +117,7 @@ contract VoteWeb3{
 
         // pushing to array
         allElections.push(electioInstance);
-        
+
         // setting start time
         setStartTime(currentElectionIndex, startTime);
 
@@ -128,32 +127,47 @@ contract VoteWeb3{
         // incrementing index for next
         currentElectionIndex++;
 
-        emit ElectionCreated(allElections[currentElectionIndex-1]);
+        emit ElectionCreated(allElections[currentElectionIndex - 1]);
     }
 
     // setting start time
-    function setStartTime(uint _electionIndex, uint amount) internal{
-        allElections[_electionIndex].startDateAndTime = block.timestamp + amount;
+    function setStartTime(uint256 _electionIndex, uint256 amount) internal {
+        allElections[_electionIndex].startDateAndTime =
+            block.timestamp +
+            amount;
         // allElections[_electionIndex].startDateAndTime = block.number + amount;
     }
 
     // setting end time
-    function setEndTime(uint _electionIndex, uint amount) internal{
+    function setEndTime(uint256 _electionIndex, uint256 amount) internal {
         allElections[_electionIndex].endDateAndTime = block.timestamp + amount;
         // allElections[_electionIndex].endDateAndTime = block.number + amount;
     }
 
-
     // Requesting Candidate
-    function requestCandidate(uint _electionIndex, address _candidateAddress, string memory _candidateName, Symbol _candidateSymbol) public{
+    function requestCandidate(
+        uint256 _electionIndex,
+        address _candidateAddress,
+        string memory _candidateName,
+        Symbol _candidateSymbol
+    ) public {
         // election should be 'Created' to add candidate (Ongoing, Finished, Discarded - should not be)
-        require(allElections[_electionIndex].electionState == ElectionState.Created, "Election state must be Created");
+        require(
+            allElections[_electionIndex].electionState == ElectionState.Created,
+            "Election state must be Created"
+        );
 
         // candidate should not be a part of election already
-        require(candidateExist[_electionIndex][_candidateAddress] == false,"You are already a Candidate in this Election");
+        require(
+            candidateExist[_electionIndex][_candidateAddress] == false,
+            "You are already a Candidate in this Election"
+        );
 
         // candidate should not have requested already
-        require(candidateRequested[_electionIndex][_candidateAddress] == false,"You have Already Requested");
+        require(
+            candidateRequested[_electionIndex][_candidateAddress] == false,
+            "You have Already Requested"
+        );
 
         // updating
         candidateRequested[_electionIndex][_candidateAddress] = true;
@@ -162,30 +176,44 @@ contract VoteWeb3{
         candiadteStructInstance.candidateAddress = _candidateAddress;
         candiadteStructInstance.candidateName = _candidateName;
         candiadteStructInstance.candidateSymbol = _candidateSymbol;
-        allElections[_electionIndex].requestedCandidates.push(candiadteStructInstance);
+        allElections[_electionIndex].requestedCandidates.push(
+            candiadteStructInstance
+        );
 
         // setting to retrive later
-        getCandidateInfo[_electionIndex][_candidateAddress] = candiadteStructInstance;
+        getCandidateInfo[_electionIndex][
+            _candidateAddress
+        ] = candiadteStructInstance;
 
         emit CandidateRequested(candiadteStructInstance);
     }
 
     // Approving Candidate
-    function approveCandidate(uint _electionIndex, address _candidateAddress) public{
+    function approveCandidate(uint256 _electionIndex, address _candidateAddress)
+        public
+    {
         // msg.sender should be manager
         require(msg.sender == manager, "You are not Authorized");
-        
+
         // election should be 'Created' to add candidate (Ongoing, Finished, Discarded - should not be)
-        require(allElections[_electionIndex].electionState == ElectionState.Created, "Election state must be Created");
+        require(
+            allElections[_electionIndex].electionState == ElectionState.Created,
+            "Election state must be Created"
+        );
 
         // candidate should not be a part of election already
-        require(candidateExist[_electionIndex][_candidateAddress] == false,"You are already a Candidate in this Election");
+        require(
+            candidateExist[_electionIndex][_candidateAddress] == false,
+            "You are already a Candidate in this Election"
+        );
 
         // updating
         candidateExist[_electionIndex][_candidateAddress] = true;
 
         // Getting candidate
-        Candidate memory getCandidate = getCandidateInfo[_electionIndex][_candidateAddress];
+        Candidate memory getCandidate = getCandidateInfo[_electionIndex][
+            _candidateAddress
+        ];
 
         // Adding candidate
         allElections[_electionIndex].approvedcandidates.push(getCandidate);
@@ -194,24 +222,33 @@ contract VoteWeb3{
     }
 
     // starting election
-    function start(uint _electionIndex) public{
+    function start(uint256 _electionIndex) public {
         // election should be 'Created' to start it (Ongoing, Finished, Discarded - should not be)
-        require(allElections[_electionIndex].electionState == ElectionState.Created, "Election state must be Created");
+        require(
+            allElections[_electionIndex].electionState == ElectionState.Created,
+            "Election state must be Created"
+        );
 
         // minimum 2 candidates required
-        require(allElections[_electionIndex].approvedcandidates.length >= 2, "Not enough Candidates");
+        require(
+            allElections[_electionIndex].approvedcandidates.length >= 2,
+            "Not enough Candidates"
+        );
 
         // updating election state
         allElections[_electionIndex].electionState = ElectionState.Ongoing;
 
         emit ElectionStarted(allElections[_electionIndex]);
-    } 
+    }
 
     // ending election
-    function end(uint _electionIndex) public{
+    function end(uint256 _electionIndex) public {
         // election should be 'Ongoing' to end it (Created, Finished, Discarded - should not be)
-        require(allElections[_electionIndex].electionState == ElectionState.Ongoing, "Election state must be Ongoing");
-        
+        require(
+            allElections[_electionIndex].electionState == ElectionState.Ongoing,
+            "Election state must be Ongoing"
+        );
+
         // updating election state
         allElections[_electionIndex].electionState = ElectionState.Finished;
 
@@ -222,24 +259,33 @@ contract VoteWeb3{
         allElections[_electionIndex].winnerVotes = highestVotes;
 
         emit ElectionEnded(allElections[_electionIndex]);
-    }   
+    }
 
     // discarding election
-    function discard(uint _electionIndex) public{
+    function discard(uint256 _electionIndex) public {
         // updating election state
         allElections[_electionIndex].electionState = ElectionState.Discarded;
-    }     
+    }
 
     // vote function
-    function vote(uint _electionIndex, address _candidateAddress) public{
+    function vote(uint256 _electionIndex, address _candidateAddress) public {
         // election should be 'Ongoing' to vote (Created, Finished, Discarded - should not be)
-        require(allElections[_electionIndex].electionState == ElectionState.Ongoing, "Election state must be Ongoing");
+        require(
+            allElections[_electionIndex].electionState == ElectionState.Ongoing,
+            "Election state must be Ongoing"
+        );
 
         // candidate can not vote
-        require(candidateExist[_electionIndex][msg.sender] == false,"Candidate can not vote");
+        require(
+            candidateExist[_electionIndex][msg.sender] == false,
+            "Candidate can not vote"
+        );
 
         // user must not have voted already
-        require(userVoted[_electionIndex][msg.sender] == false, "You have already voted");
+        require(
+            userVoted[_electionIndex][msg.sender] == false,
+            "You have already voted"
+        );
 
         // marking it true
         userVoted[_electionIndex][msg.sender] = true;
@@ -254,28 +300,34 @@ contract VoteWeb3{
         // highestVotes = max(highestVotes, candidateVotes[_electionIndex][_candidateAddress]);
 
         // updating highest votes and potential winner as people vote
-        if(highestVotes < candidateVotes[_electionIndex][_candidateAddress]){
+        if (highestVotes < candidateVotes[_electionIndex][_candidateAddress]) {
             highestVotes = candidateVotes[_electionIndex][_candidateAddress];
             currentWinningCandidate = _candidateAddress;
         }
     }
 
     // get recieved votes of a candidate
-    function getVotes(uint _electionIndex, address _candidateAddress) public returns(uint){
+    function getVotes(uint256 _electionIndex, address _candidateAddress)
+        public
+        returns (uint256)
+    {
         emit GetElection(allElections[_electionIndex]);
         return candidateVotes[_electionIndex][_candidateAddress];
     }
 
     // get single election info
-    function getElection(uint _electionIndex) public returns(Election memory){
+    function getElection(uint256 _electionIndex)
+        public
+        returns (Election memory)
+    {
         emit GetElection(allElections[_electionIndex]);
         return allElections[_electionIndex];
     }
 
     // get single election info
-    function getAllElection() public view returns(Election[] memory){
+    function getAllElection() public view returns (Election[] memory) {
         return allElections;
-    } 
+    }
 }
 
 // request candidate func
@@ -284,16 +336,16 @@ contract VoteWeb3{
 // candidate can directly apply or create account
 
 // tests to be done:
-// - manager check
-// - create election
-// - add candidate 
-// - check candidate array
-// - add same candidate again
-// - min 2 candidates needed
+//* - manager check
+//* - create election
+//* - add candidate
+//* - check candidate array
+//* - add same candidate again
+//* - min 2 candidates needed
 // - vote before starting
-// - start election and then add candidate 
+// - start election and then add candidate
 // - vote after starting
-// - vote from same account to the same election 
+// - vote from same account to the same election
 //     - to same candidate
 //     - to different candidate
 // - create another election
@@ -302,11 +354,11 @@ contract VoteWeb3{
 // - vote to candidates from different accounts
 //     - 2 votes to one candidate
 //     - 1 vote to the other candidate
-// - end first election 
+// - end first election
 //     - check winner
 // - check first election state
-// - end 2nd election 
-//     - check winner 
+// - end 2nd election
+//     - check winner
 // - check 2nd election state
 // - create another election
 // - add 2 candidates
